requirements.txt
asgiref==3.5.0
dj-rest-auth==2.2.3
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
mysql-client==0.0.1
Pillow==9.0.1
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
